feat(extractor): add Mistral OCR fallback for scanned PDFs and SSE support
- Enable Mistral OCR fallback when PDF has no extractable text
- Fix Dockerfile to install root package with OCR extras (mistralai)
- Parse Mistral date format (DD/MM/YYYY) correctly in ContaDistribuidora
- Add PDF generation support with fpdf (migration 000006)
- Add SSE real-time event broker via pg_notify
- Add consumo_minimo_kwh column to contract (migration 000007)

+++ services/backend-go/internal/billing/engine_types.go
@@ -46,7 +46,7 @@ type CalcContract struct {
 type CalculationInput struct {
 	Contract         CalcContract         `json:"contract"`
 	Itens            []UtilityInvoiceItem `json:"itens"`
-	ConsumoMinimoKWh int                  `json:"consumo_minimo_kwh"`
+	ConsumoMinimoKWh float64              `json:"consumo_minimo_kwh"`
 }
 
 // LineBreakdown é uma linha do detalhamento do resultado.
